refactor(handler): T2b.5a — relocate handler-bound fns out of src/tools
Move the two JulieServerHandler-naming functions out of the tool layer
into the top crate (handler-adjacent), making all 9 tool dirs handler-free.

New files:
- src/handler/embedding_init.rs: wait_for_embedding_provider_settled +
  supporting statics + cfg(test) tracking helpers
- src/handler/workspace_resolution.rs: resolve_workspace_filter +
  workspace_resolution_failure + suggest_closest_workspace

Trimmed files (handler-free helpers retained):
- src/tools/search/nl_embeddings.rs: keeps maybe_initialize_embeddings_for_nl_definitions;
  re-exports take_nl_definition_embedding_init_attempts from new location
- src/tools/navigation/resolution.rs: keeps parse_qualified_name,
  definition_priority, compare_symbols_by_priority_and_context,
  file_path_matches_suffix, normalize_path_suffix, WorkspaceTarget re-export

Callers updated:
- handler/tool_context_impl.rs: both backing-fn paths repointed
- handler/tools/{fast_search,fast_refs,get_context}.rs: import new path
- tests/tools/workspace/global_targeting.rs: split import
- tools/search/mod.rs: crate::health::SystemStatus → julie_core::health_types

Done-check: rg JulieServerHandler src/tools/{9 dirs} → 0 hits
Gate: cargo nextest run -p julie --no-run + cargo build → clean
