+++ .memories/2026-06-04/132045_307c.md
@@ -1,0 +1,72 @@
+---
+id: checkpoint_307c9d78
+timestamp: 2026-06-04T13:20:45.041Z
+tags:
+  - refactor
+  - julie-rescue
+  - phase2b
+  - T2b.5a
+  - de-handler
+  - embedding-init
+  - workspace-resolution
+git:
+  branch: julie-rescue-phase2b
+  commit: 3729fa3e
+  files:
+    - .codex/config.toml
+    - .miller/.symbols.db.init.lock
+    - .miller/indexer.lock
+    - .miller/logs/miller-20260531.jsonl
+    - .miller/logs/miller-20260603.jsonl
+    - .miller/logs/miller-35698-20260531.jsonl
+    - .miller/logs/miller-38072-20260531.jsonl
+    - .miller/logs/miller-43360-20260531.jsonl
+    - .miller/symbols.db
+    - .miller/symbols.db.julie-extract.lock
+    - src/handler.rs
+    - src/handler/embedding_init.rs
+    - src/handler/tool_context_impl.rs
+    - src/handler/tools/fast_refs.rs
+    - src/handler/tools/fast_search.rs
+    - src/handler/tools/get_context.rs
+    - src/handler/workspace_resolution.rs
+    - src/tests/tools/workspace/global_targeting.rs
+    - src/tools/navigation/resolution.rs
+    - src/tools/search/mod.rs
+    - src/tools/search/nl_embeddings.rs
+summary: "T2b.5a complete: de-handler the 9 tool dirs"
+briefId: julie-rescue-de-slop-packaging-keep-the-moat
+type: checkpoint
+evidence:
+  - cargo nextest run -p julie --no-run → Finished 0.25s (cached)
+  - cargo build → Finished 16.48s, 0 errors
+  - rg JulieServerHandler src/tools/{9 dirs} → exit 1 (0 matches)
+next: T2b.5b — physical julie-tools crate extraction
+---
+
+## T2b.5a complete: de-handler the 9 tool dirs
+
+Moved the two handler-bound functions out of `src/tools/` into the top crate, making all 9 tool dirs `JulieServerHandler`-free.
+
+### What moved
+
+**`wait_for_embedding_provider_settled`** → `src/handler/embedding_init.rs`
+- Took with it: `NL_DEFINITION_EMBEDDING_INIT_SINGLE_FLIGHT`, `NL_DEFINITION_EMBEDDING_INIT_ATTEMPTS` (cfg(test)), `record_nl_definition_embedding_init_attempt` (cfg(test)), `take_nl_definition_embedding_init_attempts` (cfg(test) pub(crate))
+- `nl_embeddings.rs` now contains only `maybe_initialize_embeddings_for_nl_definitions` + a cfg(test) re-export of `take_nl_definition_embedding_init_attempts` so existing test consumers keep working
+
+**`resolve_workspace_filter` + private helpers** → `src/handler/workspace_resolution.rs`
+- Took with it: `workspace_resolution_failure` (private), `suggest_closest_workspace` (private)
+- `resolution.rs` now contains only handler-free helpers: `parse_qualified_name`, `WorkspaceTarget` re-export, `definition_priority`, `compare_symbols_by_priority_and_context`, `file_path_matches_suffix`, `normalize_path_suffix`
+
+### Updated callers
+- `handler/tool_context_impl.rs` — both call paths repointed to new crate-local modules
+- `handler/tools/{get_context,fast_search,fast_refs}.rs` — import from `handler::workspace_resolution`
+- `tests/tools/workspace/global_targeting.rs` — split import; `resolve_workspace_filter` from new path
+- `tools/search/mod.rs:43` — `crate::health::SystemStatus` → `julie_core::health_types::SystemStatus`
+
+### Module registration
+Added `pub(crate) mod embedding_init` and `pub(crate) mod workspace_resolution` to `src/handler.rs` with `#[path]` attributes.
+
+### Done-check
+- `rg JulieServerHandler src/tools/{search,navigation,...}` → **0 hits**
+- Gates: `cargo nextest run -p julie --no-run` → clean; `cargo build` → 16.5s clean
